--- pkg/ui/init.go
@@ -1,9 +1,0 @@
-package ui
-
-import (
-	tea "github.com/charmbracelet/bubbletea"
-)
-
-func (m Model) Init() tea.Cmd {
-	return nil
-}
